- Bugfix; deleting a chapter will update the file - Cleaned delete module - Breadcrumb nav is sticky to the top

+++ src/networkRouting/editorPages/DeleteModule.kt
@@ -10,25 +10,25 @@ object DeleteModule : EditorModule() {
     suspend fun PipelineContext<*, ApplicationCall>.deleteActions() {
         tapeCatalog.tapes.firstOrNull { it.name == call.parameters["tape"] }
             ?.also { tape ->
-                val chapterName = call.parameters["chapter"]
-                if (chapterName == null) {
-                    tapeCatalog.tapes.remove(tape)
-                    if (tape.file?.exists().isTrue())
-                        tape.file?.delete()
-                } else {
-                    tape.chapters.removeIf { it.name == chapterName }
-
-                    call.respondRedirect {
-                        val pathStart = this.encodedPath.substringBeforeLast('/')
-                            .removePrefix("/")
-                        path(pathStart, TapeRouting.RoutePaths.EDIT.path)
-                        parameters.clear()
-                        parameters.append("tape", tape.name)
+                when (val chapterName = call.parameters["chapter"]) {
+                    null -> {
+                        if (tapeCatalog.tapes.remove(tape) && tape.file?.exists().isTrue())
+                            tape.file?.delete()
+                        call.respondRedirect(TapeRouting.RoutePaths.ALL.path)
                     }
-                    return
+                    else -> {
+                        if (tape.chapters.removeIf { it.name == chapterName } && tape.file?.exists().isTrue())
+                            tape.saveFile()
+
+                        call.respondRedirect {
+                            val pathStart = this.encodedPath.substringBeforeLast('/')
+                                .removePrefix("/")
+                            path(pathStart, TapeRouting.RoutePaths.EDIT.path)
+                            parameters.clear()
+                            parameters.append("tape", tape.name)
+                        }
+                    }
                 }
             }
-
-        call.respondRedirect(TapeRouting.RoutePaths.ALL.path)
     }
 }
